+++ cloudsmith_cli/cli/commands/push.py
@@ -13,13 +13,14 @@ from ...core.api.packages import (
     create_package as api_create_package, get_package_status,
     get_package_formats)
 from ...core import utils
+from ...core.api.exceptions import ApiException
 from .. import decorators, validators
 from ..exceptions import handle_api_exceptions
 from ..types import ExpandPath
 from . import main
 
 
-def upload_file(ctx, opts, owner, repo, filepath):
+def upload_file(ctx, opts, owner, repo, filepath, skip_errors):
     """Upload a package file via the API."""
     filename = click.format_filename(filepath)
     basename = os.path.basename(filename)
@@ -31,7 +32,9 @@ def upload_file(ctx, opts, owner, repo, filepath):
     )
 
     context_msg = "Failed to request file upload!"
-    with handle_api_exceptions(ctx, opts=opts, context_msg=context_msg):
+    with handle_api_exceptions(
+            ctx, opts=opts, context_msg=context_msg,
+            reraise_on_error=skip_errors):
         with spinner():
             identifier, upload_url, upload_fields = request_file_upload(
                 owner=owner,
@@ -70,7 +73,8 @@ def upload_file(ctx, opts, owner, repo, filepath):
 
 
 def create_package(
-        ctx, opts, owner, repo, package_file_id, package_type, **kwargs):
+        ctx, opts, owner, repo, package_file_id, package_type,
+        skip_errors, **kwargs):
     """Create a new package via the API."""
     click.echo()
     click.echo(
@@ -84,7 +88,9 @@ def create_package(
         payload.update(kwargs)
 
     context_msg = "Failed to create package!"
-    with handle_api_exceptions(ctx, opts=opts, context_msg=context_msg):
+    with handle_api_exceptions(
+            ctx, opts=opts, context_msg=context_msg,
+            reraise_on_error=skip_errors):
         with spinner():
             slug_perm, slug = api_create_package(
                 package_format=package_type,
@@ -107,7 +113,8 @@ def create_package(
     return slug_perm, slug
 
 
-def wait_for_package_sync(ctx, opts, owner, repo, slug, wait_interval):
+def wait_for_package_sync(
+        ctx, opts, owner, repo, slug, wait_interval, skip_errors):
     """Wait for a package to synchronise (or fail)."""
     click.echo()
     completed = False
@@ -129,7 +136,9 @@ def wait_for_package_sync(ctx, opts, owner, repo, slug, wait_interval):
         }
 
     context_msg = "Failed to synchronise file!"
-    with handle_api_exceptions(ctx, opts=opts, context_msg=context_msg):
+    with handle_api_exceptions(
+            ctx, opts=opts, context_msg=context_msg,
+            reraise_on_error=skip_errors):
         last_progress = 0
         with click.progressbar(
                 length=100, label=label, fill_char=fill_char,
@@ -149,11 +158,13 @@ def wait_for_package_sync(ctx, opts, owner, repo, slug, wait_interval):
         click.secho("Package synchronised successfully!", fg='green')
     else:
         click.secho("Package failed to synchronise!", fg='red')
+        click.echo("Please see the frontend for additional errors (we'll ")
+        click.echo("expose more information to the CLI here later.)")
 
 
 def upload_files_and_create_package(
         ctx, opts, package_type, package_file, owner_repo, dry_run,
-        no_wait_for_sync, wait_interval, **kwargs):
+        no_wait_for_sync, wait_interval, skip_errors, **kwargs):
     """Upload package files and create a new package."""
     owner, repo = owner_repo
 
@@ -162,7 +173,8 @@ def upload_files_and_create_package(
 
     # 2. Upload the primary package file and store the id
     package_file_id = upload_file(
-        ctx=ctx, opts=opts, owner=owner, repo=repo, filepath=package_file
+        ctx=ctx, opts=opts, owner=owner, repo=repo, filepath=package_file,
+        skip_errors=skip_errors
     )
 
     # 3. Clean additional arguments, upload any that look like files
@@ -175,14 +187,15 @@ def upload_files_and_create_package(
             continue
 
         kwargs[k] = upload_file(
-            ctx=ctx, opts=opts, owner=owner, repo=repo, filepath=v
+            ctx=ctx, opts=opts, owner=owner, repo=repo, filepath=v,
+            skip_errors=skip_errors
         )
 
     # 4. Create the package with package files and additional arguments
     slug_perm, slug = create_package(
         ctx=ctx, opts=opts, owner=owner, repo=repo,
         package_file_id=package_file_id, package_type=package_type,
-        **kwargs
+        skip_errors=skip_errors, **kwargs
     )
 
     if no_wait_for_sync:
@@ -191,7 +204,7 @@ def upload_files_and_create_package(
     # 5. (optionally) Wait for the package to synchronise
     wait_for_package_sync(
         ctx=ctx, opts=opts, owner=owner, repo=repo, slug=slug,
-        wait_interval=wait_interval
+        wait_interval=wait_interval, skip_errors=skip_errors
     )
 
 
@@ -214,17 +227,39 @@ PUSH_HANDLERS = {}  # For keeping references to the handlers
 PUSH_CONTEXT = get_package_formats()
 
 for key, parameters in six.iteritems(PUSH_CONTEXT):
+    kwargs = parameters.copy()
+
+    # Remove standard arguments
+    kwargs.pop('package_file')
+    if 'distribution' in parameters:
+        has_distribution_param = True
+        kwargs.pop('distribution')
+    else:
+        has_distribution_param = False
+
+    has_additional_params = len(kwargs) > 0
+
     help_text=(
         """
         Push/upload a new %(type)s package upstream.
-
-        - PACKAGE_FILE: The main file for the package.
         """ % {
             'type': key.capitalize()
         }
     )
 
-    if 'distribution' in parameters:
+    if has_additional_params:
+        help_text += """
+
+        - PACKAGE_FILE: The main file to create the package from.
+        """
+    else:
+        help_text += """
+
+        - PACKAGE_FILE: Any number of files to create packages from. Each file
+        will result in a separate package.
+        """
+
+    if has_distribution_param:
         target_metavar = 'OWNER/REPO/DISTRO[/VERSION]'
         target_callback = validators.validate_owner_repo_distro
         help_text += """
@@ -257,8 +292,10 @@ for key, parameters in six.iteritems(PUSH_CONTEXT):
         metavar=target_metavar,
         callback=target_callback)
     @click.argument(
-        'package_file', type=ExpandPath(
-            dir_okay=False, exists=True, writable=False, resolve_path=True
+        'package_file',
+        nargs=1 if has_additional_params else -1,
+        type=ExpandPath(
+            dir_okay=False, exists=True, writable=False, resolve_path=True,
         ))
     @click.option(
         '-n', '--dry-run', default=False, is_flag=True,
@@ -269,6 +306,9 @@ for key, parameters in six.iteritems(PUSH_CONTEXT):
     @click.option(
         '-I', '--wait-interval', default=5.0, type=float, show_default=True,
         help="The time in seconds to wait between checking operations.")
+    @click.option(
+        '-s', '--skip-errors', default=False, is_flag=True,
+        help="Skip errors when uploading multiple files.")
     @decorators.common_cli_config_options
     @decorators.common_cli_output_options
     @decorators.common_api_auth_options
@@ -292,19 +332,25 @@ for key, parameters in six.iteritems(PUSH_CONTEXT):
             owner_repo = owner_repo[0:2]
         kwargs['owner_repo'] = owner_repo
 
-        upload_files_and_create_package(ctx, *args, **kwargs)
+        package_files = kwargs.pop('package_file')
+        if not isinstance(package_files, tuple):
+            package_files = (package_files,)
 
-    kwargs = parameters.copy()
+        for package_file in package_files:
+            kwargs['package_file'] = package_file
 
-    # Remove standard arguments
-    kwargs.pop('package_file')
-    if 'distribution' in parameters:
-        kwargs.pop('distribution')
+            try:
+                upload_files_and_create_package(ctx, *args, **kwargs)
+            except ApiException:
+                assert(kwargs.get('skip_errors'))
+                click.secho("Skipping error and moving on.", fg='yellow')
+
+            click.echo()
 
     # Add any additional arguments
     for k, info in six.iteritems(kwargs):
-        # This is also a bit hacky. :-)
         if k.endswith('_file'):
+            # Treat parameters that end with _file as uploadable filepaths.
             option_type=ExpandPath(
                 dir_okay=False, exists=True, writable=False, resolve_path=True
             )