+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **Added** `api_host`, `api_proxy` and `api_user_agent` to config file.
 - **Added** `help` command for those who need more than `-h` and `--help`.
 - **Added** service endpoint to `check` command output.
+- **Added** ability to upload multiple package files at once.
 - **Changed** environment variables to use a `CLOUDSMITH_` prefix.
 - **Fixed** validation for `push` commands that require a distribution.
 - **Fixed** token endpoint failing because API key overrides login/password.